+++ ckanext/gobar_theme/routing.py
@@ -11,7 +11,7 @@ class GobArRouter:
 
         self.home_controller = 'ckanext.gobar_theme.controller:GobArHomeController'
         self.home_routes = SubMapper(self.route_map, controller=self.home_controller)
-        self.api_controller = 'ckanext.gobar_theme.controller:GobArApiController'
+        self.api_controller = 'ckanext.gobar_theme.api_controller:GobArApiController'
         self.package_controller = 'ckanext.gobar_theme.package_controller:GobArPackageController'
         self.config_controller = 'ckanext.gobar_theme.config_controller:GobArConfigController'
         self.user_controller = 'ckanext.gobar_theme.user_controller:GobArUserController'
